requirements.txt
certifi==2019.11.28
chardet==3.0.4
falcon==2.0.0
falcon-jinja2==0.0.3
gunicorn==20.0.4
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.7
